Interviewcake Trees and Graphs
Final challenge from the trees and graph section. "Find a Duplicate Beast Mode"
challenged me to find a duplicate integer value in the range of 1 -n in and array
of length n+1, in O(n) time and O(1) space. Very challenging, required using the
values at each index like a pointer for the next index in order to find a loop
in the array and eventually to find the duplicate value.